refactor: 💡 move function for handle reach
